requirements.txt
starlette
uvicorn[standard]
requests
dadaptation
wandb
pyngrok
pycloudflared
scipy
came-pytorch
pytorch_optimizer==3.1.2
#https://github.com/67372a/LyCORIS/tree/dev
../installables/lycoris_lora-3.1.0.67372a0-py3-none-any.whl
wheel
typing
triton==3.0.0; platform_system == "Linux"

# Triton wheels
https://github.com/woct0rdho/triton-windows/releases/download/v3.0.0-windows.post3/triton-3.0.0-cp312-cp312-win_amd64.whl; platform_system == "Windows" and python_version == "3.12"
https://github.com/woct0rdho/triton-windows/releases/download/v3.0.0-windows.post3/triton-3.0.0-cp311-cp311-win_amd64.whl; platform_system == "Windows" and python_version == "3.11"
https://github.com/woct0rdho/triton-windows/releases/download/v3.0.0-windows.post3/triton-3.0.0-cp310-cp310-win_amd64.whl; platform_system == "Windows" and python_version == "3.10"
